requirements.txt
beartype>=0.14.1
datasets>=2.14.5
einops>=0.7.0
ipython>=8.5.0
jaxtyping>=0.2.25
matplotlib>=3.5.0
numpy>=1.21.5
torch>=2.1.0
transformers>=4.33.3
lightning>=2.0.0
PyYAML>=5.4.1
tqdm>=4.66.1
dvc>=3.28
dvclive>=3.41.1
dvc_gdrive>=2.20.0
